# Run mypy test faster