@samuelcolvin 2020-11-29
Humm, I have no idea.

Is this really worth the effort?