@MrMrRobat 2020-11-29
I'm not sure, just was playing around to see if I could make theese tests run faster. 
Maybe it's worth opening issue in pytest repo (not necessairly for sake of solving this issue).

As for now, I don't have motivation right now to dig further, but it may change overtime. If you ok with hackiness introduced there, maybe someday I will find a fix, otherwise, you can close this PR.